#2231 - Add support for WADO / Streaming
This is interim commit.

- Refactored the code to support streaming of compressed pixel data. The pixel data is decompressed by StreamingClient before being returned to the caller. StreamingClient relies on the information returned by the server to during the decompression.

- Modified the code to deal with server partitions. The server must be able to distinguish studies that exist on more than one partitions. For eg, we may store JPEG2000 lossless studies on ServerA, and same studies but compressed using JPEG lossy on ServerB. To request for JPEG2000 images, the client will send requests to http://localhost:1000/wado/ServerA?requesttype=.... To request for the JPEG lossy version of the same images (only sop instance uid is different ), it will send requests to  http://localhost:1000/wado/ServerB?requesttype=....

- Modified the StreamingClient to accept serverAE from the caller.

- Renamed some namespaces and classes.

- Added ServerPartitionMonitor to monitor the server partition settings. It's similar to FileSystemMonitor except it's singleton. Applications subscribe to its Changed event to be notified when a partition is modified. In the future we may want to change the code in dicom server to listen to this event.

+++ ImageServer/Services/Streaming/ImageStreaming/WADORequestTypeHandlerManager.cs
@@ -66,7 +66,7 @@ namespace ClearCanvas.ImageServer.Services.Streaming.ImageStreaming
             if (_handlers.ContainsKey(requestType))
                 return _handlers[requestType];
             else
-                throw new WADOException((int)HttpStatusCode.BadRequest, String.Format("Unsupported RequestType {0}", requestType));
+                throw new WADOException(HttpStatusCode.BadRequest, String.Format("Unsupported RequestType {0}", requestType));
 
         }
